+++ services/authServices.js
@@ -61,13 +61,7 @@ export const loginUser = async ({ email, password }) => {
     };
 };
 
-export const logoutUser = async user => {
-    if (!user) {
-        throw HttpError(401, 'Not authorized');
-    }
-    user.token = null;
-    await user.save();
-};
+
 
 export const getCurrentUser = async user => {
     if (!user) {
