+++ services/userServices.js
@@ -48,6 +48,14 @@ export const loginUser = async ({ email, password }) => {
   };
 };
 
+export const logoutUser = async user => {
+  if (!user) {
+    throw HttpError(401, 'Not authorized');
+  }
+  user.token = null;
+  await user.save();
+};
+
 export const getFollowingByUserId = async userId => {
   const followingLinks = await Follow.findAll({
     where: { followerId: userId },
